backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
Pillow>=10.0.0
httpx>=0.27.0
huggingface-hub>=0.23.0

# Local VLM inference (Qwen3-VL-2B-Instruct)
torch>=2.4,<2.7
torchvision>=0.19,<0.22
transformers==5.5.3
accelerate>=0.34
qwen-vl-utils
numpy
